+++ config/config.go
@@ -34,4 +34,5 @@ func setDefaultValues() {
 	viper.SetDefault(settings.HighlightHeadlinesKey, settings.HighlightHeadlinesDefault)
 	viper.SetDefault(settings.RelativeNumberingKey, settings.RelativeNumberingDefault)
 	viper.SetDefault(settings.HideYCJobsKey, settings.HideYCJobsDefault)
+	viper.SetDefault(settings.UseAlternateIndentBlockKey, settings.UseAlternateIndentBlockDefault)
 }